Work/Хакатон от города Москва/ML-Serivce/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
confluent-kafka==2.4.0
pydantic==2.7.4
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
pyyaml==6.0.1
joblib==1.3.2
pytz==2024.1
pyarrow==16.1.0          # для parquet
SQLAlchemy==2.0.32       # если захочешь Postgres
psycopg[binary]==3.1.18  # если захочешь Postgres
